requirements.txt
pyyaml>=6.0
pytest>=7.0
loguru>=0.7
